techk/requirements.txt
attrs==19.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-filter==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
flake8==3.5.0
idna==2.6
Markdown==2.6.11
mccabe==0.6.1
more-itertools==7.2.0
pep8==1.7.1
pluggy==0.6.0
psycopg2==2.7.4
py==1.8.0
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.5.1
pytz==2018.4
requests==2.18.4
six==1.12.0
urllib3==1.22
